requirements-dev.txt
black==25.1.0
drf-spectacular==0.28.0
flake8==7.1.2
isort==6.0.1
psycopg2-binary==2.9.10
pytest-cov==5.0.0
pytest==8.3.5
pytest-django==4.10.0